requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
SQLAlchemy==2.0.36
Pillow==11.0.0
pillow-heif==0.18.0
Werkzeug==3.0.1
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1